requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
kaleido>=0.2.1
numpy>=1.26.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
